chore(call-recorder): bump SDK to 2.16 and extract CallRecordingUpdateFields type
