+++ packages/twenty-apps/public/call-recorder/src/logic-functions/data/update-call-recording.util.ts
@@ -1,24 +1,6 @@
 import { type CoreApiClient } from 'twenty-client-sdk/core';
 
-import { type CallRecordingRequestStatus } from 'src/logic-functions/constants/call-recording-request-status';
-import { type CallRecordingStatus } from 'src/logic-functions/constants/call-recording-status';
-
-export type CallRecordingUpdateFields = Partial<{
-  // null clears a previously synced title when the calendar title disappears.
-  title: string | null;
-  status: CallRecordingStatus;
-  recordingRequestStatus: CallRecordingRequestStatus;
-  startedAt: string;
-  endedAt: string;
-  calendarEventId: string;
-  // null clears stale app-owned state on cancel/eject or reschedule.
-  externalBotId: string | null;
-  externalRecordingId: string;
-  callRecorderFailureReason: string | null;
-  transcript: Record<string, unknown>;
-  audio: { fileId: string; label: string }[];
-  video: { fileId: string; label: string }[];
-}>;
+import { type CallRecordingUpdateFields } from 'src/logic-functions/types/call-recording-update-fields.type';
 
 export const updateCallRecording = async (
   client: CoreApiClient,
