+++ packages/twenty-apps/public/call-recorder/src/logic-functions/flows/reconcile-call-recording-transcript-artifact-result.type.ts
@@ -1,4 +1,4 @@
-import { type CallRecordingUpdateFields } from 'src/logic-functions/data/update-call-recording.util';
+import { type CallRecordingUpdateFields } from 'src/logic-functions/types/call-recording-update-fields.type';
 
 type CallRecordingTranscriptArtifactUpdateFields = Pick<
   CallRecordingUpdateFields,
