+++ packages/twenty-apps/public/call-recorder/src/logic-functions/types/call-recording-media-file.type.ts
@@ -1,0 +1,1 @@
+export type CallRecordingMediaFile = { fileId: string; label: string };
